+++ exif/__init__.py
@@ -1,4 +1,6 @@
 """Read and modify image EXIF metadata using Python."""
 
-from exif._constants import ColorSpace, Orientation, Saturation, SensingMethod, Sharpness
+from exif._constants import (
+    ColorSpace, ExposureMode, ExposureProgram, Orientation, Saturation, SceneCaptureType, SensingMethod, Sharpness,
+    WhiteBalance)
 from exif._image import Image